+++ ewm-main-service/src/main/java/ru/practicum/model/category/Category.java
@@ -16,7 +16,7 @@ public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    long id;
+    int id;
 
     @Column(name = "category_name", nullable = false, unique = true)
     String name;
